Create Panagram_Soution.java
This program is a solution to finding panagrams within a single line of text. Panagrams are made up instances where every letter from the alphabet appears at least once in a line of text. The program reads input from a .txt file and outputs whether your line of text contains a panagram or not to the console. For the input, put the number of lines in the form of an integer on the first line of input, then type out your lines of text. My program is capable of handling up to 3 maximum panagrams in a single line of text.